pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.40.3",
    "gspread>=6.2.1",
    "python-telegram-bot==20.8",
    "telegram>=0.0.1",
]
